refactor(instrumentation/httptrace): extract default query obfuscator to obfuscator.go
